+++ hub-client/src/pages/Admin.vue
@@ -177,10 +177,6 @@
 		if (await dialog.okcancel(t('admin.remove_room_sure'))) {
 			try {
 				await rooms.removePublicRoom(room.room_id);
-				if (secured.value) {
-					// If the room was secured, we need to remove it from allowed_to_join_room table
-					rooms.kickUsersFromSecuredRoom(room.room_id);
-				}
 			} catch (error) {
 				dialog.confirm('ERROR', error as string);
 			}
@@ -192,6 +188,8 @@
 		if (await dialog.okcancel(t('admin.secured_remove_sure'))) {
 			try {
 				await rooms.removeSecuredRoom(room);
+				// If the room was secured, we need to remove the members from the allowed_to_join_room table
+				rooms.kickUsersFromSecuredRoom(room.room_id);
 			} catch (error) {
 				dialog.confirm('ERROR', error as string);
 			}